remove widgetContent
