--- src/sql/parts/dashboard/contents/widgetContent.component.html
@@ -1,10 +1,0 @@
-<!--
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the Source EULA. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
--->
-<div [ngGrid]="gridConfig" *ngIf="widgets" >
-	<dashboard-widget-wrapper *ngFor="let widget of widgets" [(ngGridItem)]="widget.gridItemConfig" [_config]="widget">
-	</dashboard-widget-wrapper>
-</div>
