--- src/sql/parts/dashboard/contents/widgetContent.css
@@ -1,9 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the Source EULA. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-widget-content {
-	height: 100%;
-	width: 100%;
-}
